Added some styling
Modified layout and added some styling
Modified file names
Moved certain files to packages
