+++ .idea/libraries/lib.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="lib">
+    <CLASSES>
+      <root url="file://$USER_HOME$/.javafx-sdk-16/lib" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES>
+      <root url="file://$USER_HOME$/.javafx-sdk-16/lib" />
+    </SOURCES>
+    <jarDirectory url="file://$USER_HOME$/.javafx-sdk-16/lib" recursive="false" />
+    <jarDirectory url="file://$USER_HOME$/.javafx-sdk-16/lib" recursive="false" type="SOURCES" />
+  </library>
+</component>
