+++ .idea/libraries/oop_project.xml
@@ -1,0 +1,9 @@
+<component name="libraryTable">
+  <library name="oop project">
+    <CLASSES>
+      <root url="jar://$PROJECT_DIR$/oop project.zip!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
