--- Main.java
@@ -1,29 +1,0 @@
-package application;
-	
-import javafx.application.Application;
-import javafx.stage.Stage;
-import javafx.scene.Scene;
-import javafx.scene.layout.BorderPane;
-import javafx.scene.layout.Pane;
-import javafx.fxml.FXMLLoader;
-
-
-public class Main extends Application {
-	@Override
-	public void start(Stage primaryStage) {
-		try {
-			Pane root = FXMLLoader.load(getClass().getResource("Sample.fxml"));
-			Scene scene = new Scene(root,900,900);
-			scene.getStylesheets().add(getClass().getResource("application.css").toExternalForm());
-			primaryStage.setScene(scene);
-			primaryStage.setTitle("PIZZA ORDER");
-			primaryStage.show();
-		} catch(Exception e) {
-			e.printStackTrace();
-		}
-	}
-	
-	public static void main(String[] args) {
-		launch(args);
-	}
-}
