+++ za/ac/tut/appStyling.css
@@ -1,0 +1,68 @@
+/* CSS styling for OOPas Pizza Shop */
+
+#appPane{
+    -fx-background-color: #34495e;
+    -fx-font-family: Arial;
+
+    -fx-border-color: #ecf0f1;
+}
+
+#topPane{
+    -fx-background-color: #34495e;
+}
+
+#appTitle{
+    -fx-font-family: Arial;
+    -fx-font-size: 20;
+    -fx-font-weight: bolder;
+    -fx-text-fill: #ababab;
+}
+
+#leftPane{
+    -fx-background-color: #2c3e50;
+}
+
+#sectionTitle{
+    -fx-text-fill: #ababab;
+}
+
+#cboBox{
+    -fx-color: #2ecc71;
+}
+
+#rdoSmall, #rdoMedium, #rdoLarge{
+    -fx-text-fill: #bdc3c7;
+    -fx-color: #27ae60;
+}
+#chkExtraMeat, #chkBlackOlives, #chkExtraCheese, #chkSausage, #ckhBBQ, #chkMushroom, #chkOnions, #chkPeriperi{
+    -fx-text-fill: #bdc3c7;
+    -fx-color: #2980b9;
+}
+
+#rdoDelivery, #rdoDineIn, #rdoTakeOut{
+    -fx-text-fill: #bdc3c7;
+    -fx-color: #3498db;
+}
+
+#btnDecrement, #btnIncrement{
+    -fx-color: #f39c12;
+}
+
+#btnBillOut, #btnOrderAgain{
+    -fx-color: #f1c40f;
+}
+
+#txtQuantity{
+    -fx-text-fill: #bdc3c7;
+}
+
+#bottomPane{
+    -fx-background-color: #34495e;
+}
+
+#footer{
+    -fx-font-family: Arial;
+    -fx-font-size: 16;
+    -fx-font-weight: bolder;
+    -fx-fill: #ababab;
+}
